Fix: hand rolled directory walking with walkdir crate use BTreeMap instead of HashMap Extract unit test from cargo_fmt file into an integration test suite Make various functions public so they can be accessible in the integration file
