+++ Cargo.toml
@@ -50,6 +50,7 @@ tracing-subscriber = { version = "0.3", features = ["fmt", "env-filter"] }
 bytes = "1.11.1"
 dialoguer = "0.12.0"
 toml_edit = "0.25.11"
+walkdir = "2.5.0"
 
 
 [dev-dependencies]
