+++ src/lib.rs
@@ -10,11 +10,11 @@ pub mod init;
 pub mod mcp;
 pub mod output;
 pub mod plugins;
+pub mod session_state;
 pub mod start;
 pub mod sync;
 
 pub(crate) mod crate_sources;
 pub(crate) mod predicate;
-pub(crate) mod session_state;
 pub(crate) mod skills;
 pub(crate) mod workspace;
